crawler/requirements.txt
PyYAML==6.0.1
SQLAlchemy==1.4.49
numpy==1.25.2
beautifulsoup4==4.12.2
requests==2.31.0
mysql-connector-python==8.1.0
html5lib==1.1
pika==1.3.2